+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.1.4 (unreleased)
 
 - Fix DLQMetadata decoding to use DLQMetadataDecoder func instead of inferred decoding from TopicType.
 - Fix consumer to use noopDLQ if RetryQ or DLQ in config is empty.
+- Fix ResetOffset fails on partition rebalance.
 
 
 v0.1.3 (2018-03-09)